refactor(ghas): Move location model class to `odg.model`
